gen-ai/agents/chatbots/jarvis/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis"
version = "0.1.0"
description = "Jarvis - a personal AI assistant for myself"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Girish", email = "rgirish1994@gmail.com"}
]
dependencies = [
    "strands-agents>=1.0.0",
    "strands-agents[ollama]",
    "strands-agents-tools>=0.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
]

[project.scripts]
jarvis = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"
